requirements.txt
python-dotenv>=1.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
streamlit>=1.30,<2.0
pandas>=2.0,<3.0
plotly>=5.18,<6.0
pytest>=7.0,<9.0
streamlit-aggrid>=1.0.5
